+++ app/src/main/java/dev/ranga/recipeexplorer/data/local/FavouriteRecipeEntity.kt
@@ -6,9 +6,9 @@ const val FAVOURITE_RECIPE_TABLE_NAME = "favourite_recipe_table"
 
 @Entity(tableName = FAVOURITE_RECIPE_TABLE_NAME, primaryKeys = ["id"])
 data class FavouriteRecipeEntity(
-    val id: Int,
+    val id: Long,
     val name: String,
     val thumbnailUrl: String,
-    val totalTimeMinutes: Int,
+    val totalTimeMinutes: Long,
     val userRatings: Double,
 )
